+++ src/Entity/RecipeProduct.php
@@ -79,7 +79,7 @@ class RecipeProduct
      * @param Recipe $recipe
      * @return $this Implementing fluent interface.
      */
-    public function setRecipe(Recipe $recipe)
+    public function setRecipe(Recipe $recipe): self
     {
         $this->recipe = $recipe;
         return $this;
@@ -99,7 +99,7 @@ class RecipeProduct
      * @param Item $item
      * @return $this Implementing fluent interface.
      */
-    public function setItem(Item $item)
+    public function setItem(Item $item): self
     {
         $this->item = $item;
         return $this;
@@ -119,7 +119,7 @@ class RecipeProduct
      * @param float $amountMin
      * @return $this Implementing fluent interface.
      */
-    public function setAmountMin(float $amountMin)
+    public function setAmountMin(float $amountMin): self
     {
         $this->amountMin = (int) ($amountMin * self::FACTOR_AMOUNT_MIN);
         return $this;
@@ -139,7 +139,7 @@ class RecipeProduct
      * @param float $amountMax
      * @return $this Implementing fluent interface.
      */
-    public function setAmountMax(float $amountMax)
+    public function setAmountMax(float $amountMax): self
     {
         $this->amountMax = (int) ($amountMax * self::FACTOR_AMOUNT_MAX);
         return $this;
@@ -159,7 +159,7 @@ class RecipeProduct
      * @param float $probability
      * @return $this Implementing fluent interface.
      */
-    public function setProbability(float $probability)
+    public function setProbability(float $probability): self
     {
         $this->probability = (int) ($probability * self::FACTOR_AMOUNT_PROBABILITY);
         return $this;
@@ -188,7 +188,7 @@ class RecipeProduct
      * @param int $order
      * @return $this Implementing fluent interface.
      */
-    public function setOrder(int $order)
+    public function setOrder(int $order): self
     {
         $this->order = $order;
         return $this;